requirement.txt
streamlit
openai
pytubefix
langchain
langchain-openai
fastmcp
pandas